# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns/{campaignId}/proofs:
    post:
      operationId: submitCampaignProof
      tags: [fulfillment]
      summary: Approve one submitted operator proof revision
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProofApprovalInput"
      responses:
        "201":
          description: Proof submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proof"
        "404": { description: Campaign not found or not authorized }
        "409": { description: Submitted proof revision is no longer available }
  /operator/campaigns/{campaignId}/placement-orders/{placementOrderId}/proofs/request-url:
    post:
      operationId: requestOperatorProductionProofUpload
      tags: [fulfillment, storage]
      summary: Issue an operator-bound production proof upload intent
      security: [{ operatorBearer: [] }]
      parameters:
        - $ref: "#/components/parameters/CampaignId"
        - $ref: "#/components/parameters/PlacementOrderId"
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ProofUploadRequest" } } }
      responses:
        "201": { description: Upload intent issued, content: { application/json: { schema: { $ref: "#/components/schemas/ProofUploadIntent" } } } }
        "401": { description: Operator authentication required }
        "409": { description: Placement is not funded }
  /operator/campaigns/{campaignId}/placement-orders/{placementOrderId}/proofs/{intentId}/finalize:
    post:
      operationId: finalizeOperatorProductionProofUpload
      tags: [fulfillment, storage]
      summary: Verify and finalize an operator proof upload once
      security: [{ operatorBearer: [] }]
      parameters:
        - $ref: "#/components/parameters/CampaignId"
        - $ref: "#/components/parameters/PlacementOrderId"
        - $ref: "#/components/parameters/UploadIntentId"
      responses:
        "200": { description: Upload intent finalized, content: { application/json: { schema: { $ref: "#/components/schemas/ProofUploadIntent" } } } }
        "401": { description: Operator authentication required }
        "409": { description: Intent expired or already finalized }
  /operator/campaigns/{campaignId}/proofs:
    post:
      operationId: submitOperatorCampaignProof
      tags: [fulfillment]
      summary: Submit a finalized proof revision for one funded placement
      security: [{ operatorBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/CampaignId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ProofSubmissionInput" } } }
      responses:
        "201": { description: Proof revision submitted, content: { application/json: { schema: { $ref: "#/components/schemas/Proof" } } } }
        "401": { description: Operator authentication required }
        "409": { description: Intent or proof revision is no longer available }
  /campaigns/{campaignId}/checkins:
    post:
      operationId: submitCampaignCheckin
      tags: [fulfillment]
      summary: Submit an owner check-in
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinSubmissionInput"
      responses:
        "201":
          description: Check-in submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkin"
        "409": { description: Check-in is not currently due }
  /campaigns/{campaignId}/checkins/photo/request-url:
    post:
      operationId: requestCampaignCheckinPhotoUpload
      tags: [fulfillment, storage]
      summary: Issue an owner-bound photo upload intent for the current check-in cycle
      parameters: [{ $ref: "#/components/parameters/CampaignId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CheckinPhotoUploadRequest" } } }
      responses:
        "201": { description: Upload intent issued, content: { application/json: { schema: { $ref: "#/components/schemas/CheckinPhotoUploadIntent" } } } }
        "409": { description: Check-in is not currently due }
  /campaigns/{campaignId}/checkins/photo/{intentId}/finalize:
    post:
      operationId: finalizeCampaignCheckinPhotoUpload
      tags: [fulfillment, storage]
      summary: Verify and finalize an owner check-in photo upload once
      parameters:
        - $ref: "#/components/parameters/CampaignId"
        - $ref: "#/components/parameters/UploadIntentId"
      responses:
        "200": { description: Upload intent finalized, content: { application/json: { schema: { $ref: "#/components/schemas/CheckinPhotoUploadIntent" } } } }
        "409": { description: Intent expired or already finalized }
  /campaigns/{campaignId}/make-good:
    post:
      operationId: selectCampaignMakeGood
      tags: [fulfillment]
      summary: Select a make-good after a missed check-in
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MakeGoodInput"
      responses:
        "200": { description: Make-good selected }
        "409": { description: A check-in has not been missed }
...
  /operator/campaigns/{campaignId}/shipment:
    post:
      operationId: recordCampaignShipment
      tags: [fulfillment]
      summary: Record shipment after proof approval
      security: [{ operatorBearer: [] }]
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipmentInput"
      responses:
        "200": { description: Shipment recorded }
        "401": { description: Operator authentication required }
        "409": { description: Proof approval or shipping address required }
...
components:
  securitySchemes:
    operatorBearer:
      type: http
      scheme: bearer
      bearerFormat: API token
  parameters:
    CampaignId:
      in: path
      name: campaignId
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
...
    PlacementOrderId:
      in: path
      name: placementOrderId
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
    UploadIntentId:
      in: path
      name: intentId
      required: true
      schema: { type: string, minLength: 1 }
  schemas:
...
    ProofSubmissionInput:
      type: object
      additionalProperties: false
      required: [placementOrderId, intentId]
      properties:
        placementOrderId: { type: string, minLength: 1 }
        intentId: { type: string, minLength: 1 }
    ProofApprovalInput:
      type: object
      additionalProperties: false
      required: [placementOrderId, revision]
      properties:
        placementOrderId: { type: string, minLength: 1 }
        revision: { type: integer, minimum: 1 }
    ProofUploadRequest:
      type: object
      additionalProperties: false
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        size: { type: integer, minimum: 1, maximum: 10000000 }
        contentType: { type: string, enum: [image/png, image/jpeg, image/webp] }
    ProofUploadIntent:
      type: object
      required: [id, purpose, objectPath, expectedMimeType, expectedSizeBytes, expectedFileName, status, expiresAt]
      properties:
        id: { type: string }
        purpose: { type: string, const: operator_production_proof }
        objectPath: { type: string, pattern: "^/objects/uploads/[A-Za-z0-9-]+$" }
        expectedMimeType: { type: string, enum: [image/png, image/jpeg, image/webp] }
        expectedSizeBytes: { type: integer }
        expectedFileName: { type: string }
        status: { type: string, enum: [issued, finalized, consumed, revoked] }
        expiresAt: { type: string, format: date-time }
        uploadURL: { type: string, format: uri }
    Proof:
      type: object
      required: [revision, objectPath, status, submittedAt]
      properties:
        revision: { type: integer, minimum: 1 }
        objectPath: { type: string }
        status: { type: string, enum: [submitted, approved, rejected] }
        submittedAt: { type: string, format: date-time }
        approvedAt: { type: [string, "null"], format: date-time }
    ShipmentInput:
      type: object
      additionalProperties: false
      required: [carrier, trackingNumber]
      properties:
        carrier: { type: string, minLength: 2, maxLength: 80 }
        trackingNumber: { type: string, minLength: 4, maxLength: 100, pattern: "^[A-Za-z0-9 ._-]+$" }
...
    CheckinSubmissionInput:
      type: object
      additionalProperties: false
      required: [note]
      properties:
        note: { type: string, minLength: 1, maxLength: 2000 }
        photoIntentId: { type: string, minLength: 1 }
    CheckinPhotoUploadRequest:
      type: object
      additionalProperties: false
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        size: { type: integer, minimum: 1, maximum: 25000000 }
        contentType: { type: string, enum: [image/png, image/jpeg, image/webp] }
    CheckinPhotoUploadIntent:
      type: object
      required: [id, purpose, objectPath, expectedMimeType, expectedSizeBytes, expectedFileName, status, expiresAt]
      properties:
        id: { type: string }
        purpose: { type: string, const: owner_checkin }
        objectPath: { type: string, pattern: "^/objects/uploads/[A-Za-z0-9-]+$" }
        expectedMimeType: { type: string, enum: [image/png, image/jpeg, image/webp] }
        expectedSizeBytes: { type: integer }
        expectedFileName: { type: string }
        status: { type: string, enum: [issued, finalized, consumed, revoked] }
        expiresAt: { type: string, format: date-time }
        uploadURL: { type: string, format: uri }
    Checkin:
      type: object
      required: [id, campaignId, note, submittedAt]
      properties:
        id: { type: string }
        campaignId: { type: string }
        note: { type: string }
        photoObjectPath: { type: [string, "null"] }
        submittedAt: { type: string, format: date-time }
    MakeGoodInput:
      type: object
      additionalProperties: false
      required: [selection]
      properties:
        selection: { type: string, enum: [extension, credit, cancellation] }