# lib/api-zod/src/generated/types/checkinPhotoUploadIntent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckinPhotoUploadIntentExpectedMimeType } from './checkinPhotoUploadIntentExpectedMimeType';
import type { CheckinPhotoUploadIntentStatus } from './checkinPhotoUploadIntentStatus';

export interface CheckinPhotoUploadIntent {
  id: string;
  purpose: 'owner_checkin';
  /** @pattern ^/objects/uploads/[A-Za-z0-9-]+$ */
  objectPath: string;
  expectedMimeType: CheckinPhotoUploadIntentExpectedMimeType;
  expectedSizeBytes: number;
  expectedFileName: string;
  status: CheckinPhotoUploadIntentStatus;
  expiresAt: Date;
  uploadURL?: string;
}